packages/moss-doc-parser/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moss-doc-parser"
version = "0.1.0"
description = "Document parsing utilities for Moss semantic search"
readme = "README.md"
license-files = ["LICENSE"]
authors = [
    { name = "InferEdge Inc.", email = "contact@moss.dev" }
]
keywords = ["search", "semantic", "document", "parser", "moss"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.10"
dependencies = [
    "pypdf>=3.0",
    "python-docx>=1.0",
    "python-pptx>=0.6",
    "beautifulsoup4>=4.12",
    "python-magic>=0.4",
    "markdown>=3.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.0.0",
    "flake8>=7.0.0",
    "mypy>=1.0.0",
    "build>=1.0.0",
    "twine>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true